requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
aioredis==2.0.1
python-multipart==0.0.6
PyPDF2==3.0.1
reportlab==4.0.7
Pillow==10.4.0
psutil
pydantic